requirements-development.txt
# Copyright (C) 2025 iolaire mcfadden.
# This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

# Development Dependencies for Vedfolnir Docker Environment

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.3.0

# Code Quality and Linting
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0
pre-commit>=3.3.0

# Debugging and Profiling
debugpy>=1.6.7
ipython>=8.14.0
ipdb>=0.13.13
memory-profiler>=0.61.0
py-spy>=0.3.14
line-profiler>=4.1.0

# Development Tools
watchdog>=3.0.0
python-dotenv>=1.0.0
rich>=13.5.0
click>=8.1.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0

# Database Development Tools
alembic>=1.11.0
sqlalchemy-utils>=0.41.0

# HTTP and API Testing
httpx>=0.24.0
requests-mock>=1.11.0
responses>=0.23.0

# Mock and Fixtures
factory-boy>=3.3.0
faker>=19.3.0
freezegun>=1.2.0

# Performance Testing
locust>=2.16.0

# Security Testing
bandit>=1.7.5
safety>=2.3.0

# Container Development
docker>=6.1.0
docker-compose>=1.29.0

# Hot Reloading and File Watching
watchfiles>=0.19.0